kernel: support for booting with cpio-based boot_archive
In addition to booting with ufs and hsfs boot archives, this commit adds
support for cpio boot archives.  The archive itself can be compressed (since
loader decompresses the image before handing it to the kernel), but the
individual files must *not* be compressed.  Size-wise, the cpio archives
tend to be marginally smaller than the hsfs archives with identical files.